restructure templates files and devide logic/style
