+++ client/src/components/templates/modals/friend/Style.js
@@ -1,0 +1,16 @@
+import styled from 'styled-components';
+
+export const CardLayout = styled.div`
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+  flex-wrap: wrap;
+  @media screen and (max-width: 860px) {
+    overflow-y: scroll;
+    -ms-overflow-style: none;
+    scrollbar-width: none;
+    &::-webkit-scrollbar {
+      display: none;
+    }
+  }
+`;
