+++ client/src/components/templates/modals/letter/Letter.js
@@ -1,7 +1,7 @@
 import { useState } from 'react';
-import Letters from '../../module/letter/Organism/Letters';
-import LetterCreate from '../../module/letter/create/LetterCreate';
-import Overlay from '../../atoms/overlay/Overlay';
+import Letters from '../../../module/letter/Organism/Letters';
+import LetterCreate from '../../../module/letter/create/LetterCreate';
+import Overlay from '../../../atoms/overlay/Overlay';
 
 const Letter = ({ pointRefresher }) => {
   const [isOpen, setIsOpen] = useState(false);
