+++ client/src/components/templates/modals/theme/ThemeStore.js
@@ -1,17 +1,17 @@
 import { useState, useEffect } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
-import { getCookie } from '../../../utils/cookie';
+import { getCookie } from '../../../../utils/cookie';
 import {
   memberIdSelector,
   paletteCodeSelector,
   myPaletteSelector,
-} from '../../../redux/hooks';
-import { setMyPalette, setPaletteCode } from '../../../redux/slice';
-import { BuyPalette, SetPalette } from '../../../api/ThemeStoreApi';
-import Button from '../../atoms/button/commonButton/Button';
-import Overlay from '../../atoms/overlay/Overlay';
-import { StoreModal } from '../../module/modal/Modal';
-import CircleCarousel from '../../module/theme/CircleCarousel';
+} from '../../../../redux/hooks';
+import { setMyPalette, setPaletteCode } from '../../../../redux/slice';
+import { BuyPalette, SetPalette } from '../../../../api/ThemeStoreApi';
+import Button from '../../../atoms/button/commonButton/Button';
+import Overlay from '../../../atoms/overlay/Overlay';
+import { StoreModal } from '../../../module/modal/Modal';
+import CircleCarousel from '../../../module/theme/CircleCarousel';
 import { toast } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
@@ -19,54 +19,7 @@ import {
   faChevronRight,
   faChevronLeft,
 } from '@fortawesome/free-solid-svg-icons';
-import styled from 'styled-components';
-
-const TitleContainer = styled.div`
-  margin: 5px;
-  height: 140px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-`;
-
-const Point = styled.div`
-  height: 50px;
-  margin-top: 20px;
-  font-size: 17px;
-  font-weight: 500;
-  color: #fcc403;
-`;
-
-const PaletteName = styled.div`
-  margin: 10px;
-  font-weight: 700;
-  font-size: 20px;
-  letter-spacing: 3px;
-`;
-
-const BtnContainer = styled.div`
-  margin: 5px;
-  display: flex;
-  justify-content: space-between;
-`;
-
-const ArrowContainer = styled.div`
-  display: flex;
-  justify-content: space-between;
-  padding-left: 45px;
-  padding-right: 45px;
-  .arrow {
-    color: gray;
-  }
-`;
-
-const CarouselContainer = styled.div`
-  position: relative;
-  width: 100%;
-  height: 100%;
-  display: block;
-  overflow: hidden;
-`;
+import * as Style from './Style';
 
 export const ThemeStore = ({ pointRefresher }) => {
   const dispatch = useDispatch();
@@ -178,14 +131,16 @@ export const ThemeStore = ({ pointRefresher }) => {
   return (
     <StoreModal>
       {popup && <Overlay />}
-      <TitleContainer>
+      <Style.TitleContainer>
         {Array.isArray(paletteInfo) && (
           <>
-            <Point>{`${paletteInfo[carouselIndex].palettePrice}P`}</Point>
-            <PaletteName>{paletteInfo[carouselIndex].paletteName}</PaletteName>
+            <Style.Point>{`${paletteInfo[carouselIndex].palettePrice}P`}</Style.Point>
+            <Style.PaletteName>
+              {paletteInfo[carouselIndex].paletteName}
+            </Style.PaletteName>
           </>
         )}
-        <BtnContainer>
+        <Style.BtnContainer>
           <Button
             size="long"
             fontsize="middle"
@@ -202,9 +157,9 @@ export const ThemeStore = ({ pointRefresher }) => {
           >
             적용
           </Button>
-        </BtnContainer>
-      </TitleContainer>
-      <ArrowContainer>
+        </Style.BtnContainer>
+      </Style.TitleContainer>
+      <Style.ArrowContainer>
         <FontAwesomeIcon
           icon={faChevronLeft}
           size="2x"
@@ -217,14 +172,14 @@ export const ThemeStore = ({ pointRefresher }) => {
           onClick={() => toRight()}
           style={{ cursor: 'pointer' }}
         />
-      </ArrowContainer>
-      <CarouselContainer>
+      </Style.ArrowContainer>
+      <Style.CarouselContainer>
         <CircleCarousel
           carouselIndex={carouselIndex}
           paletteInfo={paletteInfo}
           setPaletteInfo={setPaletteInfo}
         />
-      </CarouselContainer>
+      </Style.CarouselContainer>
     </StoreModal>
   );
 };
