+++ client/src/components/templates/pages/home/Style.js
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+export const Browser = styled.div`
+  position: relative;
+  max-width: 1140px;
+  margin: 0 auto;
+  padding: 0 20px;
+`;
