+++ client/src/components/templates/pages/login/Login.js
@@ -1,7 +1,7 @@
 import { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { useNavigate } from 'react-router-dom';
-import GoogleLogin from '../../module/header/login/GoogleLogin';
+import GoogleLogin from '../../../module/header/login/GoogleLogin';
 
 export default function Login() {
   return <GoogleLogin text="로그인" />;
