+++ client/src/components/templates/pages/logout/Logout.js
@@ -1,7 +1,7 @@
 import { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { useNavigate } from 'react-router-dom';
-import { LogoutApi } from '../../../api/LoginLogoutApi';
+import { LogoutApi } from '../../../../api/LoginLogoutApi';
 import GoogleLogin from '../module/GoogleLogin';
 
 export default function Logout() {
