+++ client/src/components/templates/pages/mobileAlert/Style.js
@@ -1,0 +1,20 @@
+import styled from 'styled-components';
+
+export const MobilePopUp = styled.div`
+  div {
+    z-index: 6;
+  }
+`;
+export const PopUp = styled.div`
+  position: absolute;
+  width: 100%;
+  z-index: 7;
+  top: 0;
+  left: 0;
+  text-align: center;
+`;
+export const LogoBox = styled.div`
+  width: 30%;
+  margin: 0 auto;
+  padding-bottom: 16px;
+`;
